[mhlo:linalg] Refactor to pass along an adaptor instead of a range of args
This lets us pass along attributes, which resolves the spacial case for compare
and reduceprecision. It also allows us to use attributes without extending this
interface in the future.

Adapt users of the interface to pass in the attributes or make up some
attributes if the input isn't actually mhlo.

PiperOrigin-RevId: 475883448