[tf:profiler] Add ScopedAnnotationStack for clients that can't rely on RAII
In XLA runtime we generate calls to tracing/profiling from the JIT compiled XLA programs, and we have to explicitly start and end activities. TraceMe already provides an API for that. This change ports that API to ScopedAnnotation.

In LLVM/MLIR the pattern that we want to codegen looks like this:

%0 = call @xla.trace.start("foo");
%1 = call @xla.gemm() -> tensor<?x?xf32>
call @xla.trace.end(%0)

PiperOrigin-RevId: 476219092