Allow the DCT op to handle dynamic sequence lengths.
The sequence length is computed using an `or` between the last dimension of the static shape and the dynamic shape, which implies that dynamic sequence lengths were intended to be supported. However, the truncating/padding logic needed to be adjusted to actually support dynamic sequence lengths.

Because the DCT type 4 calls the DCT type 2 with a transform length `n` of 2 times the sequence length, this implies that a dynamic value of `n` also has to be supported internally (but is only supported internally and not in the public API).

Additionally, the `idct` function no longer validates the arguments twice.

PiperOrigin-RevId: 476398289