Make ResourceVariable a subclass of CompositeTensor.
By doing so, we will be able to nest ResourceVariables in ExtensionType/CompositeTensor. It will also clear our way to return ResourceVariables from tf.function and pass variable through control flow.

Note, with this CL, tf.nest with expand_composites=True will not expand resource variables into dt_resource tensors. That will be done in a following CL to reduce the chance of being rolled back for this CL.

`CompositeTensor` is added to baseclasses of `tf.experimental.dtensor.DVariable` because it is a subclass of ResourceVariable.

PiperOrigin-RevId: 476416307