Fix the error report in padded_batch.
This is raising a type error because it's calling the type:

```
TypeError: __init__() missing 3 required positional arguments: 'op', 'value_index', and 'dtype'
```

And reporting the `value_type` here is not helpful because there are cases where the you'll have a RaggedTensorSpec but the value_type is `Tensor`.

So report the spec.

PiperOrigin-RevId: 477777930