Disable quantization for unsupported cases in the XLA path
If a composite function is quantized but couldn't be converted to XLA ops, it will remain in i32xi32 form; leading to a bad performance. This cl makes sure all quantized composite functions can be converted to XLA ops by adding checks in the LiftQuantizableSpotsAsFunctionsPass.

For XLA path, unsupported cases includes:
- DepthwiseConv (the quantized DepthwiseConv has much worse performance than non-quantized one)
- Non-constant rhs
- Conv2D with unknown channel dimension

PiperOrigin-RevId: 478938728