Fix crash in tf.nn.conv2d_transpose when output shape is invalid
This PR tries to address the issue in tf.nn.conv2d_transpose where
crash will happen when output shape is invalid.

This PR adds a warpper ShapeFromFormatWithStatus (previously ShapeFromFormat)
so that it is possible to return status.

Note the change can be applied to other places to address similar crashs.
Will create follow up PRs to cover other places once this PR is merged.

This PR fixes 57958.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>