Added `SparseTensor.set_shape` to update the static dense shape of a sparse tensor.
`sparse_reorder` now preserves the static rank and statically known dimensions of the tensor.

Previously, `sparse_reorder` would only keep the static rank and dimensions if the static shape was fully defined, and would otherwise make the shape fully dynamic. There is however value in keeping a static rank and know static dimensions even if the shape is not fully defined. One common use case is when the batch size is not known but other dimensions are fully defined.

PiperOrigin-RevId: 479108317