[xla:gml_st] Avoid crashing when calling ForOp's verifier
Sometimes ForOp::verify() is called in the midst of ForOp's creation (e.g. when trying to print something with --debug).
In such situations, we may want to tolerate some incomplete state of the operation, since it is not fully created, by relaxing some assumptions.
This patch avoids the verifier to crash when trying to access the destination operands of ForOp's terminator when these are not yet updated.

PiperOrigin-RevId: 482419982