Fix histogram_fixed_width segfault.
There were two cases: one when the bincount is 1 and the range overflows, in which case
all inputs will be put into the one-and-only-bin.  The other is when the range overflows
but there is more than one bin, so the `step` will actually be finite.  Here we modify
the step computation to avoid the overflow.

Fixes #57896.

PiperOrigin-RevId: 482549975