Add slice_index property to JAX GpuDevice.
This allow mesh_utils.create_hybrid_device_mesh to distinguish
between devices attached to the same fast network e.g., NVLink
on the same host vs. across host.

We use the notion of a slice to denotate a domain of fast network,
for GPUs, this may mean a host where local GPUs are connected by
NVLink.

PiperOrigin-RevId: 485447323