Partially infer shape for TF ShapeOp for further shape inference
In cases where the Shape op's operand is not static, the TF shape inference pass doesn't infer individual elements that are required for shape inference in users of the value.

For example, in the newly added test, Shape op's result is used in Reshape op with operand of type tensor<3x7xf32>. So, if we know all but one element of the shape, we could infer shape for the Reshape op as number of elements are known.

PiperOrigin-RevId: 487645538